back-end/requirements.txt
boto3==1.34.82
channels==4.1.0
Django==5.0.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
pydantic==2.7.2
python-dotenv==1.0.1
pytz==2024.1
gunicorn==21.2.0
openai==1.43.0
pillow==10.3.0
psycopg2==2.9.9
whitenoise==6.6.0
